feat(dashboard): create pie chart component
Finishing the dashboard page styles in the styling fase of this product's
development, the last component needed was a pie chart to display the
pizzashop products popularity.

This commit uses a different chart from the Recharts library, the pie chart,
which helps the user in viewing the product popularity in an easier and
intuitive way. It will receive data from the api such as the line chart,
created a few commits before this one, and the data will be shown as a child
component of Recharts' pie chart component.