feat: create 404 not found page
In order to prevent users from accessing a non existing route and find a
blank page, a page is needed to tell the user the page does not exist and
provide a way back to the application.

This commit completes the initial styling fase of this products development
by adding a minimalist page that tells the user the page was not found
and displaying a button that allows the user to go back to the app's main
route.